pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "speconn-runtime-python"
version = "0.3.0"
description = "Python runtime for Speconn RPC"
requires-python = ">=3.11"
dependencies = ["specodec-runtime-python>=0.0.2", "anyio>=4.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/speconn"]

[project.optional-dependencies]
pyreqwest = ["pyreqwest"]
httpx = ["httpx"]

[tool.pytest.ini_options]
anyio_mode = "auto"